Formatting multimedia file types for recognition
To tell different types of multimedia files (image, videos, or sounds), the client is now
required to specify an extra extension for the filename ('IMG', 'VID', or 'SND').

When certain files are requested, the type will be sent back in the HTTP headers.
Futhermore, if the GET request has a header of image only statement, the multimedia file
will be sent only if it is an image ('IMG').

Refer to README.md for details.