Initializing database with json configs
Previously the database tables are directly manipulated in mysql
interactive consoles. Now for better view move the table configs
to a json file, and each time the server starts it create the tables
if not already existing.

This should make looking up table specifics as well as modifying it
easier.